Update login process.
The login() function was too large and the authentication conditions
were too obsfucated.

I've spilt the function into smaller named portions and rewritten the
authentication conditions. A new option has been added to turn off the
MPPass authentication (it is port specific).

This gives space for a new modes where both MPPass and passwd
authentication are turned on or off.  In the "on" mode the user has the
*option* to add an alternative login with /pass and /setpass, but only
if they are currently logged in.  In the "off" mode no auth is required,
but again a password can be set if wanted.

Also, add logging of which auth method is used.